Split out the push rule evaluator into a separate file so it can be more readily reused. Should be functionally identical.
